test: (raft) extract common driver utilities into test/pylib
